Files restructured from component to pages
