+++ src/pages/admin/AllUser.jsx
@@ -12,13 +12,12 @@ import 'aos/dist/aos.css';
 import 'tiny-slider/dist/tiny-slider.css';
 import AOS from 'aos';
 import axios from 'axios';
-import { AdminNavbar } from './AdminNavbar';
+import { AdminNavbar } from '../../components/admin/AdminNavbar';
+import { Bounce, toast } from 'react-toastify';
 
 
-export const AllInquiry = () => {
-
-    const [inquiry, setinquiry] = useState([])
-    
+export const AllUser = () => {
+    const [user, setuser] = useState([])
 
     useEffect(() => {
         AOS.init({
@@ -27,16 +26,41 @@ export const AllInquiry = () => {
           once: true,
           offset: 120
         });
-        getAllInquiry()
+        getAllUsers()
       }, []);
 
-      const getAllInquiry = async () => {
-        const res = await axios.get("/inquiry/get")
+      const getAllUsers = async () => {
+        const res = await axios.get("/user")
         console.log(res.data.data);
-        setinquiry(res.data.data)
+        setuser(res.data.data)
 
         
       }
+
+
+      const deleteUser = async (id) => {
+        const res = await axios.delete(`/user/${id}`)
+        console.log(res);
+        if (res.status == 204) {
+          toast.success('🦄 User deleted Succesfully', {
+            position: "top-left",
+            autoClose: 3000,
+            hideProgressBar: false,
+            closeOnClick: false,
+            pauseOnHover: true,
+            draggable: true,
+            progress: undefined,
+            theme: "dark",
+            transition: Bounce,
+            });
+        }
+        getAllUsers()
+         
+        
+        
+      }
+
+
   return (
     <div>
         <AdminNavbar/>
@@ -48,7 +72,7 @@ export const AllInquiry = () => {
                   <div className="row justify-content-center align-items-center">
                     <div className="col-lg-9 text-center mt-5">
                       <h1 className="heading" data-aos="fade-up" data-aos-delay="200">
-                        All Inquiry
+                        All Users
                       </h1>
                       
                     </div>
@@ -61,28 +85,33 @@ export const AllInquiry = () => {
                 <thead>
                     <tr>
                         <td>ID</td>
-                        <th>Fullname</th>
+                        <th>Firstname</th>
+                        <th>Lastname</th>
                         <th>Email</th>
-                        <th>Phone Number</th>
-                        <th>Message</th>
-                     
+                        <th>Phone No.</th>
+                        <th>Gender</th>
+                        <th>Delete</th>
+                        
                         
                     </tr>
                 </thead>
                 <tbody>
-                {inquiry?.map((inquiry)=>{
+                {user?.map((user)=>{
                     return(
                         <tr>
-                            <td>{inquiry._id}</td>
-                            <td>{inquiry.fullName}</td>
+                            <td>{user._id}</td>
+                            <td>{user.firstName}</td>
+                            <td>{user.lastName}</td>
+                            <td>{user.email}</td>
+                            <td>{user.contact}</td>
+                            <td>{user.gender}</td>
+                            <td><button onClick={() => {deleteUser(user._id)}}  className='btn btn-danger '>Delete</button></td>
                             
-                            <td>{inquiry.email}</td>
-                            <td>{inquiry.phoneNumber}</td>
-                            <td>{inquiry.message}</td>
                             
                         </tr>
                     )
                 })}
+
                 </tbody>
 
             </table>
