+++ src/pages/public/About.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react'
-import { Navbar } from './Navbar'
+import { Navbar } from '../../components/common/Navbar'
 import '../../landing/assets/fonts/icommon/style.css';
 import '../../landing/assets/fonts/flaticon/flaticon.css';
 import '../../landing/css/tiny-slider.css';
@@ -19,7 +19,7 @@ import heroBg2 from '../../landing/assets/img/hero_bg_2.jpg';
 import img1 from '../../landing/assets/img/img_1.jpg';
 import img2 from '../../landing/assets/img/img_1.jpg';
 import img3 from '../../landing/assets/img/img_3.jpg';
-import { Footer } from './Footer';
+import { Footer } from '../../components/common/Footer';
 
 
 export const About = () => {
