+++ src/pages/public/ContactUs.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from 'react'
-import { Navbar } from './Navbar'
+import { Navbar } from '../../components/common/Navbar'
 import '../../landing/assets/fonts/icommon/style.css';
 import '../../landing/assets/fonts/flaticon/flaticon.css';
 import '../../landing/css/tiny-slider.css';
@@ -15,7 +15,7 @@ import 'tiny-slider/dist/tiny-slider.css';
 import AOS from 'aos';
 
 import heroBg1 from '../../landing/assets/img/hero_bg_1.jpg';
-import { Footer } from './Footer';
+import { Footer } from '../../components/common/Footer';
 
 
 export const ContactUs = () => {
