+++ src/pages/public/PropertySingle.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react'
-import { Navbar } from './Navbar'
+import { Navbar } from '../../components/common/Navbar'
 import '../../landing/assets/fonts/icommon/style.css';
 import '../../landing/assets/fonts/flaticon/flaticon.css';
 import '../../landing/css/tiny-slider.css';
@@ -13,14 +13,14 @@ import '../../landing/js/tiny-slider';
 import 'aos/dist/aos.css';
 import 'tiny-slider/dist/tiny-slider.css';
 import AOS from 'aos';
-import { Footer } from './Footer';
+import { Footer } from '../../components/common/Footer';
 import heroBg3 from '../../landing/assets/img/hero_bg_3.jpg';
 import agent from '../../landing/assets/img/person_5-min.jpg';
 import img2 from '../../landing/assets/img/img_2.jpg';
 import { Link, useParams } from 'react-router-dom';
 import axios from 'axios';
 
-export const PropertSingle = () => {
+export const PropertySingle = () => {
 
   const [property, setProperty] = useState({})
   const [favouriteProperty, setFavouriteProperty] = useState([])
