+++ src/utils/index.js
@@ -1,0 +1,56 @@
+// Utility functions for the real estate finder application
+
+/**
+ * Format price with commas and currency symbol
+ * @param {number} price - The price to format
+ * @param {string} currency - Currency symbol (default: '$')
+ * @returns {string} Formatted price
+ */
+export const formatPrice = (price, currency = '$') => {
+  if (!price) return 'Price on request';
+  return `${currency}${price.toLocaleString()}`;
+};
+
+/**
+ * Format date to readable string
+ * @param {string|Date} date - Date to format
+ * @returns {string} Formatted date
+ */
+export const formatDate = (date) => {
+  if (!date) return '';
+  const dateObj = new Date(date);
+  return dateObj.toLocaleDateString('en-US', {
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric'
+  });
+};
+
+/**
+ * Truncate text to specified length
+ * @param {string} text - Text to truncate
+ * @param {number} maxLength - Maximum length
+ * @returns {string} Truncated text
+ */
+export const truncateText = (text, maxLength = 100) => {
+  if (!text || text.length <= maxLength) return text;
+  return text.substring(0, maxLength) + '...';
+};
+
+/**
+ * Validate email format
+ * @param {string} email - Email to validate
+ * @returns {boolean} Whether email is valid
+ */
+export const isValidEmail = (email) => {
+  const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+  return emailRegex.test(email);
+};
+
+/**
+ * Generate random ID
+ * @returns {string} Random ID
+ */
+export const generateId = () => {
+  return Math.random().toString(36).substr(2, 9);
+};
